bitmanipulation

+++ 0x14-bit_manipulation/100-get_endianness.c
@@ -1,5 +1,3 @@
-
-  
 #include "main.h"
 
 /**
@@ -14,5 +12,6 @@ int get_endianness(void)
 
 	x = 1;
 	y = (char *)&x;
-	return (*y);
+	return ((int)*y);
 }
+
